+++ CHANGELOG.md
@@ -1,2 +1,12 @@
-## 2020/04/05 fork from 
- - https://github.com/cweijan/vscode-autohotkey
+## 2020/04/05 fork from
+
+- <https://github.com/cweijan/vscode-autohotkey>
+
+## 2020/04/12 let gotoDefinition usually < 10ms
+
+```
+interface docMap {
+  key: string;
+  obj: vscode.SymbolInformation[] | null
+}
+```